Move TF1IE to its own file
