+++ youtube_dl/extractor/wat.py
@@ -1,5 +1,3 @@
-# coding: utf-8
-
 import json
 import re
 
@@ -48,31 +46,3 @@ class WatIE(InfoExtractor):
                 'title': title,
                 'thumbnail': thumbnail,
                 }
-
-class TF1IE(InfoExtractor):
-    """
-    TF1 uses the wat.tv player, currently it can only download videos with the
-    html5 player enabled, it cannot download HD videos or the news.
-    """
-    _VALID_URL = r'http://videos.tf1.fr/.*-(.*?).html'
-    _TEST = {
-        u'url': u'http://videos.tf1.fr/auto-moto/citroen-grand-c4-picasso-2013-presentation-officielle-8062060.html',
-        u'file': u'6bysb.mp4',
-        u'md5': u'66789d3e91278d332f75e1feb7aea327',
-        u'info_dict': {
-            u"title": u"Citroën Grand C4 Picasso 2013 : présentation officielle"
-        }
-    }
-
-    def _real_extract(self, url):
-        mobj = re.match(self._VALID_URL, url)
-        id = mobj.group(1)
-        webpage = self._download_webpage(url, id)
-        embed_url = self._html_search_regex(r'"(https://www.wat.tv/embedframe/.*?)"',
-                                webpage, 'embed url')
-        embed_page = self._download_webpage(embed_url, id, u'Downloading embed player page')
-        wat_id = self._search_regex(r'UVID=(.*?)&', embed_page, 'wat id')
-        wat_info = self._download_webpage('http://www.wat.tv/interface/contentv3/%s' % wat_id, id, u'Downloading Wat info')
-        wat_info = json.loads(wat_info)['media']
-        wat_url = wat_info['url']
-        return self.url_result(wat_url, 'Wat')
